requirements.txt
asgiref==3.8.1
asttokens==3.0.0
crispy-tailwind==1.0.3
decorator==5.2.1
Django==5.2
django-crispy-forms==2.4
djangorestframework==3.16.0
executing==2.2.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
